+++ frontend/src/components/Input.tsx
@@ -1,22 +1,10 @@
 import { useRef, useState, KeyboardEvent, MouseEvent, ChangeEvent } from "react";
 
-import FormControl from '@mui/material/FormControl';
 import FileUploadIcon from "@mui/icons-material/FileUpload";
-import FileOpenIcon from '@mui/icons-material/FileOpen';
 import SendIcon from "@mui/icons-material/Send";
-import Alert from '@mui/material/Alert';
-import Dialog from '@mui/material/Dialog';
-import DialogTitle from '@mui/material/DialogTitle';
-import DialogContent from '@mui/material/DialogContent';
-import TextField from '@mui/material/TextField';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemAvatar from '@mui/material/ListItemAvatar';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemText from '@mui/material/ListItemText';
-import Avatar from '@mui/material/Avatar';
 import TextareaAutosize from "react-textarea-autosize";
 import Config from "../config";
+import Foundry from "./Foundry";
 import "./Input.css";
 import { MessageDict } from "../App"
 
@@ -26,24 +14,13 @@ export default function Input(props: {
   onSendMessage: (msg: string) => void,
   onStartUpload: (filename: string) => void,
   onCompletedUpload: (msg: string) => void,
-  foundryFolder: string | undefined,
-  onSelectFoundryFolder: (folder: string) => void,
-  foundryAvailableDatasets: { name: string; dataset_rid: string; }[] | undefined;
   onSelectFoundryDataset: (name: string) => void; }
 ) {
 
   const fileInputRef = useRef<HTMLInputElement>(null);
   const [inputIsFocused, setInputIsFocused] = useState<boolean>(false);
   const [userInput, setUserInput] = useState<string>('');
-  const [foundryDialogOpen, setFoundryDialogOpen] = useState(false);
   const [messageReplay, setMessageReplay] = useState(0);
-
-  const handleFoundryDialogClose = (value: string) => {
-    setFoundryDialogOpen(false);
-    if (value != '') {
-      props.onSelectFoundryDataset(value);
-    }
-  };
 
   const handleUpload = (e: MouseEvent) => {
     e.preventDefault();
@@ -122,47 +99,10 @@ export default function Input(props: {
       <div className={"input-holder " + (inputIsFocused ? "focused" : "")}>
 
         <form className="file-open">
-          <button type="button" onClick={ () => setFoundryDialogOpen(true) }>
-            <FileOpenIcon />
-          </button>
+          <Foundry
+            onSelectFoundryDataset = { props.onSelectFoundryDataset }
+          />
         </form>
-
-        <Dialog open={foundryDialogOpen} onClose={() => handleFoundryDialogClose('') } fullWidth maxWidth="sm">
-          <DialogTitle>Select Foundry Dataset</DialogTitle>
-          <DialogContent sx={{ display: 'flex', flexDirection: 'column' }}>
-          <FormControl fullWidth>
-            <TextField
-              margin="dense"
-              id="foundry-folder"
-              label="Folder (RID or Path))"
-              value={props.foundryFolder}
-              onChange={e => { props.onSelectFoundryFolder(e.target.value); }}
-            />
-          </FormControl>
-            {props.foundryAvailableDatasets != undefined && props.foundryAvailableDatasets.length > 0 &&
-              <List sx={{ pt: 0 }}>
-                {props.foundryAvailableDatasets.map((dataset) => (
-                  <ListItem disableGutters key={dataset.dataset_rid}>
-                    <ListItemButton onClick={() => handleFoundryDialogClose(dataset.dataset_rid)}>
-                      <ListItemAvatar>
-                        <Avatar>
-                          <FileOpenIcon />
-                        </Avatar>
-                      </ListItemAvatar>
-                      <ListItemText primary={dataset.name} />
-                    </ListItemButton>
-                  </ListItem>
-                ))}
-              </List>
-            }
-            {props.foundryAvailableDatasets != undefined && props.foundryAvailableDatasets.length == 0 &&
-              <Alert severity="warning">No datasets found in folder</Alert>
-            }
-            {props.foundryAvailableDatasets == undefined &&
-              <Alert severity="error">Access to folder failed. Does it exist? Do you have access permissions?</Alert>
-            }
-          </DialogContent>
-        </Dialog>
 
         <form className="file-upload">
           <input