+++ frontend/public/assets/Changelog.md
@@ -1,4 +1,14 @@
 # Changelog #
+## March 19, 2024 ##
+* Also store selected options to recall them when re-opening the web page
+* Ensure there is always a default Foundry folder selected
+* Set default Foundry folder to `/Global/Foundry Training and Resources/Example Data/Aviation Ontology` as this should be readable by everybody
+* Added support for folder navigationi n Foundry dialog
+
+## March 18, 2024 ##
+* Added option to hide the generated code
+* Added option to attempt automatically fixing runtime errors
+* Added a button for printing the chat history. This should also allow for conveninet pdf export through your browser's print dialog.
 
 ## March 06, 2024 ##
 * Added this changelog to the documentation section